+++ chain/beacon/drand/drand.go
@@ -5,12 +5,14 @@ import (
 	"context"
 	"time"
 
-	dchain "github.com/drand/drand/chain"
-	dclient "github.com/drand/drand/client"
-	hclient "github.com/drand/drand/client/http"
-	dcrypto "github.com/drand/drand/crypto"
-	dlog "github.com/drand/drand/log"
-	gclient "github.com/drand/drand/lp2p/client"
+	dcommon "github.com/drand/drand/v2/common"
+	dchain "github.com/drand/drand/v2/common/chain"
+	dlog "github.com/drand/drand/v2/common/log"
+	dcrypto "github.com/drand/drand/v2/crypto"
+	dclient "github.com/drand/go-clients/client"
+	hclient "github.com/drand/go-clients/client/http"
+	gclient "github.com/drand/go-clients/client/lp2p"
+	drand "github.com/drand/go-clients/drand"
 	"github.com/drand/kyber"
 	lru "github.com/hashicorp/golang-lru/v2"
 	logging "github.com/ipfs/go-log/v2"
@@ -39,7 +41,7 @@ var log = logging.Logger("drand")
 // The root trust for the Drand chain is configured from buildconstants.DrandConfigs
 type DrandBeacon struct {
 	isChained bool
-	client    dclient.Client
+	client    drand.Client
 
 	pubkey kyber.Point
 
@@ -92,15 +94,14 @@ func NewDrandBeacon(genesisTs, interval uint64, ps *pubsub.PubSub, config dtypes
 		return nil, xerrors.Errorf("unable to unmarshal drand chain info: %w", err)
 	}
 
-	var clients []dclient.Client
+	var clients []drand.Client
 	for _, url := range config.Servers {
-		hc, err := hclient.NewWithInfo(url, drandChain, nil)
+		hc, err := hclient.NewWithInfo(&logger{&log.SugaredLogger}, url, drandChain, nil)
 		if err != nil {
 			return nil, xerrors.Errorf("could not create http drand client: %w", err)
 		}
-		hc.(DrandHTTPClient).SetUserAgent("drand-client-lotus/" + build.NodeBuildVersion)
+		hc.SetUserAgent("drand-client-lotus/" + build.NodeBuildVersion)
 		clients = append(clients, hc)
-
 	}
 
 	opts := []dclient.Option{
@@ -112,7 +113,11 @@ func NewDrandBeacon(genesisTs, interval uint64, ps *pubsub.PubSub, config dtypes
 	if ps != nil {
 		opts = append(opts, gclient.WithPubsub(ps))
 	} else {
-		log.Info("drand beacon without pubsub")
+		if len(clients) == 0 {
+			// This is necessary to convince a drand beacon to start without any clients. For historical
+			// beacons we need them to be able to verify old entries but we don't need to fetch new ones.
+			clients = append(clients, dclient.EmptyClientWithInfo(drandChain))
+		}
 	}
 
 	client, err := dclient.Wrap(clients, opts...)
@@ -131,7 +136,7 @@ func NewDrandBeacon(genesisTs, interval uint64, ps *pubsub.PubSub, config dtypes
 		localCache: lc,
 	}
 
-	sch, err := dcrypto.GetSchemeByIDWithDefault(drandChain.Scheme)
+	sch, err := dcrypto.GetSchemeByID(drandChain.Scheme)
 	if err != nil {
 		return nil, err
 	}
@@ -165,8 +170,8 @@ func (db *DrandBeacon) Entry(ctx context.Context, round uint64) <-chan beacon.Re
 		if err != nil {
 			br.Err = xerrors.Errorf("drand failed Get request: %w", err)
 		} else {
-			br.Entry.Round = resp.Round()
-			br.Entry.Data = resp.Signature()
+			br.Entry.Round = resp.GetRound()
+			br.Entry.Data = resp.GetSignature()
 		}
 		log.Debugw("done fetching randomness", "round", round, "took", build.Clock.Since(start))
 		out <- br
@@ -192,7 +197,7 @@ func (db *DrandBeacon) VerifyEntry(entry types.BeaconEntry, prevEntrySig []byte)
 		// return no error if the value is in the cache already
 		return nil
 	}
-	b := &dchain.Beacon{
+	b := &dcommon.Beacon{
 		PreviousSig: prevEntrySig,
 		Round:       entry.Round,
 		Signature:   entry.Data,
@@ -239,10 +244,10 @@ var _ beacon.RandomBeacon = (*DrandBeacon)(nil)
 
 func BeaconScheduleFromDrandSchedule(dcs dtypes.DrandSchedule, genesisTime uint64, ps *pubsub.PubSub) (beacon.Schedule, error) {
 	shd := beacon.Schedule{}
-	for _, dc := range dcs {
+	for i, dc := range dcs {
 		bc, err := NewDrandBeacon(genesisTime, buildconstants.BlockDelaySecs, ps, dc.Config)
 		if err != nil {
-			return nil, xerrors.Errorf("creating drand beacon: %w", err)
+			return nil, xerrors.Errorf("creating drand beacon #%d: %w", i, err)
 		}
 		shd = append(shd, beacon.BeaconPoint{Start: dc.Start, Beacon: bc})
 	}